⭕(wip): refactor shipment form
